Adding support for "account" and "key" settings when creating an "azure" snapshot repo.
