+++ src/main/java/org/elasticsearch/cloud/azure/storage/AzureStorageServiceImpl.java
@@ -20,29 +20,17 @@
 package org.elasticsearch.cloud.azure.storage;
 
 import com.microsoft.azure.storage.CloudStorageAccount;
-import com.microsoft.azure.storage.StorageException;
-import com.microsoft.azure.storage.blob.*;
+import com.microsoft.azure.storage.blob.CloudBlobClient;
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.cloud.azure.AzureSettingsFilter;
-import org.elasticsearch.common.blobstore.BlobMetaData;
-import org.elasticsearch.common.blobstore.support.PlainBlobMetaData;
-import org.elasticsearch.common.collect.ImmutableMap;
 import org.elasticsearch.common.component.AbstractLifecycleComponent;
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.settings.SettingsFilter;
 import org.elasticsearch.common.unit.TimeValue;
-import org.elasticsearch.repositories.RepositoryException;
-
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.net.URI;
-import java.net.URISyntaxException;
 
 import static org.elasticsearch.cloud.azure.storage.AzureStorageService.Storage.ACCOUNT;
-import static org.elasticsearch.cloud.azure.storage.AzureStorageService.Storage.ACCOUNT_DEPRECATED;
 import static org.elasticsearch.cloud.azure.storage.AzureStorageService.Storage.KEY;
-import static org.elasticsearch.cloud.azure.storage.AzureStorageService.Storage.KEY_DEPRECATED;
 
 /**
  *
@@ -50,22 +38,27 @@ import static org.elasticsearch.cloud.azure.storage.AzureStorageService.Storage.
 public class AzureStorageServiceImpl extends AbstractLifecycleComponent<AzureStorageServiceImpl>
     implements AzureStorageService {
 
-    private final String account;
-    private final String key;
-    private final String blob;
-    private final TimeValue timeout;
-
-    private CloudBlobClient client;
-
     @Inject
     public AzureStorageServiceImpl(Settings settings, SettingsFilter settingsFilter) {
         super(settings);
         settingsFilter.addFilter(new AzureSettingsFilter());
-        // We try to load storage API settings from `cloud.azure.`
-        account = settings.get(ACCOUNT, settings.get(ACCOUNT_DEPRECATED));
-        key = settings.get(KEY, settings.get(KEY_DEPRECATED));
-        timeout = settings.getAsTime(Storage.TIMEOUT, TimeValue.timeValueMinutes(-1));
-        blob = "https://" + account + ".blob.core.windows.net/";
+    }
+
+    @Override
+    public synchronized AzureClientImpl client() {
+
+        String account = componentSettings.get("account", settings.get(ACCOUNT));
+        String key = componentSettings.get("key", settings.get(KEY));
+
+        return client(account, key);
+    }
+
+    @Override
+    public synchronized AzureClientImpl client(String account, String key) {
+
+        CloudBlobClient cloudBlobClient = null;
+        TimeValue timeout = settings.getAsTime(Storage.TIMEOUT, TimeValue.timeValueMinutes(-1));
+        String blob = "https://" + account + ".blob.core.windows.net/";
 
         try {
             if (account != null) {
@@ -80,17 +73,19 @@ public class AzureStorageServiceImpl extends AbstractLifecycleComponent<AzureSto
                 CloudStorageAccount storageAccount = CloudStorageAccount.parse(storageConnectionString);
 
                 // Create the blob client.
-                client = storageAccount.createCloudBlobClient();
+                cloudBlobClient = storageAccount.createCloudBlobClient();
 
                 // Set timeout option. See cloud.azure.storage.timeout or cloud.azure.storage.xxx.timeout
                 if (timeout.getSeconds() > 0) {
-                    client.getDefaultRequestOptions().setTimeoutIntervalInMs(safeTimeInMsToInt(timeout));
+                    cloudBlobClient.getDefaultRequestOptions().setTimeoutIntervalInMs(safeTimeInMsToInt(timeout));
                 }
             }
         } catch (Exception e) {
             // Can not start Azure Storage Client
             logger.error("can not start azure storage client: {}", e.getMessage());
         }
+
+        return new AzureClientImpl(settings, cloudBlobClient);
     }
 
     // Needed for Java < Java 8
@@ -100,147 +95,6 @@ public class AzureStorageServiceImpl extends AbstractLifecycleComponent<AzureSto
                     "]. It can not be longer than 2,147,483,647ms.");
         }
         return (int) timeout.getMillis();
-    }
-
-    @Override
-    public boolean doesContainerExist(String container) {
-        try {
-            CloudBlobContainer blobContainer = client.getContainerReference(container);
-            return blobContainer.exists();
-        } catch (Exception e) {
-            logger.error("can not access container [{}]", container);
-        }
-        return false;
-    }
-
-    @Override
-    public void removeContainer(String container) throws URISyntaxException, StorageException {
-        CloudBlobContainer blobContainer = client.getContainerReference(container);
-        // TODO Should we set some timeout and retry options?
-        /*
-        BlobRequestOptions options = new BlobRequestOptions();
-        options.setTimeoutIntervalInMs(1000);
-        options.setRetryPolicyFactory(new RetryNoRetry());
-        blobContainer.deleteIfExists(options, null);
-        */
-        logger.trace("removing container [{}]", container);
-        blobContainer.deleteIfExists();
-    }
-
-    @Override
-    public void createContainer(String container) throws URISyntaxException, StorageException {
-        try {
-            CloudBlobContainer blobContainer = client.getContainerReference(container);
-            logger.trace("creating container [{}]", container);
-            blobContainer.createIfNotExists();
-        } catch (IllegalArgumentException e) {
-            logger.trace("fails creating container [{}]", container, e.getMessage());
-            throw new RepositoryException(container, e.getMessage());
-        }
-    }
-
-    @Override
-    public void deleteFiles(String container, String path) throws URISyntaxException, StorageException {
-        logger.trace("delete files container [{}], path [{}]", container, path);
-
-        // Container name must be lower case.
-        CloudBlobContainer blobContainer = client.getContainerReference(container);
-        if (blobContainer.exists()) {
-            // We list the blobs using a flat blob listing mode
-            for (ListBlobItem blobItem : blobContainer.listBlobs(path, true)) {
-                String blobName = blobNameFromUri(blobItem.getUri());
-                logger.trace("removing blob [{}] full URI was [{}]", blobName, blobItem.getUri());
-                deleteBlob(container, blobName);
-            }
-        }
-    }
-
-    /**
-     * Extract the blob name from a URI like https://myservice.azure.net/container/path/to/myfile
-     * It should remove the container part (first part of the path) and gives path/to/myfile
-     * @param uri URI to parse
-     * @return The blob name relative to the container
-     */
-    public static String blobNameFromUri(URI uri) {
-        String path = uri.getPath();
-
-        // We remove the container name from the path
-        // The 3 magic number cames from the fact if path is /container/path/to/myfile
-        // First occurrence is empty "/"
-        // Second occurrence is "container
-        // Last part contains "path/to/myfile" which is what we want to get
-        String[] splits = path.split("/", 3);
-
-        // We return the remaining end of the string
-        return splits[2];
-    }
-
-    @Override
-    public boolean blobExists(String container, String blob) throws URISyntaxException, StorageException {
-        // Container name must be lower case.
-        CloudBlobContainer blobContainer = client.getContainerReference(container);
-        if (blobContainer.exists()) {
-            CloudBlockBlob azureBlob = blobContainer.getBlockBlobReference(blob);
-            return azureBlob.exists();
-        }
-
-        return false;
-    }
-
-    @Override
-    public void deleteBlob(String container, String blob) throws URISyntaxException, StorageException {
-        logger.trace("delete blob for container [{}], blob [{}]", container, blob);
-
-        // Container name must be lower case.
-        CloudBlobContainer blobContainer = client.getContainerReference(container);
-        if (blobContainer.exists()) {
-            logger.trace("container [{}]: blob [{}] found. removing.", container, blob);
-            CloudBlockBlob azureBlob = blobContainer.getBlockBlobReference(blob);
-            azureBlob.delete();
-        }
-    }
-
-    @Override
-    public InputStream getInputStream(String container, String blob) throws URISyntaxException, StorageException {
-        logger.trace("reading container [{}], blob [{}]", container, blob);
-        return client.getContainerReference(container).getBlockBlobReference(blob).openInputStream();
-    }
-
-    @Override
-    public OutputStream getOutputStream(String container, String blob) throws URISyntaxException, StorageException {
-        logger.trace("writing container [{}], blob [{}]", container, blob);
-        return client.getContainerReference(container).getBlockBlobReference(blob).openOutputStream();
-    }
-
-    @Override
-    public ImmutableMap<String, BlobMetaData> listBlobsByPrefix(String container, String keyPath, String prefix) throws URISyntaxException, StorageException {
-        logger.debug("listing container [{}], keyPath [{}], prefix [{}]", container, keyPath, prefix);
-        ImmutableMap.Builder<String, BlobMetaData> blobsBuilder = ImmutableMap.builder();
-
-        CloudBlobContainer blobContainer = client.getContainerReference(container);
-        if (blobContainer.exists()) {
-            for (ListBlobItem blobItem : blobContainer.listBlobs(keyPath + (prefix == null ? "" : prefix))) {
-                URI uri = blobItem.getUri();
-                logger.trace("blob url [{}]", uri);
-
-                // uri.getPath is of the form /container/keyPath.* and we want to strip off the /container/
-                // this requires 1 + container.length() + 1, with each 1 corresponding to one of the /
-                String blobPath = uri.getPath().substring(1 + container.length() + 1);
-
-                CloudBlockBlob blob = blobContainer.getBlockBlobReference(blobPath);
-
-                // fetch the blob attributes from Azure (getBlockBlobReference does not do this)
-                // this is needed to retrieve the blob length (among other metadata) from Azure Storage
-                blob.downloadAttributes();
-
-                BlobProperties properties = blob.getProperties();
-                String name = blobPath.substring(keyPath.length());
-                logger.trace("blob url [{}], name [{}], size [{}]", uri, name, properties.getLength());
-                blobsBuilder.put(name, new PlainBlobMetaData(name, properties.getLength()));
-            }
-        }
-
-        return blobsBuilder.build();
     }
 
     @Override
