+++ src/test/java/org/elasticsearch/repositories/azure/AzureSnapshotRestoreITest.java
@@ -483,7 +483,7 @@ public class AzureSnapshotRestoreITest extends AbstractAzureTest {
         assertThat(awaitBusy(new Predicate<Object>() {
             public boolean apply(Object obj) {
                 try {
-                    storageService.createContainer(container);
+                    storageService.client().createContainer(container);
                     logger.debug(" -> container created...");
                     return true;
                 } catch (URISyntaxException e) {
@@ -496,7 +496,7 @@ public class AzureSnapshotRestoreITest extends AbstractAzureTest {
                 }
             }
         }, 30, TimeUnit.SECONDS), equalTo(true));
-        storageService.removeContainer(container);
+        storageService.client().removeContainer(container);
 
         ClusterAdminClient client = client().admin().cluster();
         logger.info("-->  creating azure repository while container is being removed");
@@ -538,7 +538,7 @@ public class AzureSnapshotRestoreITest extends AbstractAzureTest {
 
         AzureStorageService client = new AzureStorageServiceImpl(settings, settingsFilter);
         for (String container : containers) {
-            client.removeContainer(container);
+            client.client().removeContainer(container);
         }
     }
 }
